[tutor_67] - Persistência de Chat (Longa Duração)
