+++ backend/application/socket/Impl/validacao_emit.py
@@ -1,21 +1,5 @@
 from datetime import datetime
-from flask import request
-from flask_socketio import SocketIO, emit
-from application.config.vector_database import collection
-from application.services.service_chat import criar_chat
-from application.services.service_mensagem import criar_mensagem, buscar_ultimas_n_mensagens
-from application.constants import LLM_UUID
 import uuid
-import ollama
-from application.models.model_chat import Chat
-from application.config.database import db
-from application.models.model_mensagem import Mensagem
-
-socketio = SocketIO(cors_allowed_origins="*", async_mode="eventlet")
-
-@socketio.on("connect")
-def handle_connect():
-    emit("connection-confirmation", {"data": "Conexão estabelecida"})
 
 def _is_valid_uuid_(value: any): 
     try:
@@ -73,53 +57,3 @@ def validacao_emit(json_emit: dict[str, any]):
         "Valido": len(erros) == 0,
         "Invalido": erros
     }
-    
-def registrar_mensagem(id_chat, usuario_id, sessao_id, data_de_envio, conteudo):
-    try:
-        if not conteudo:
-            print(f"[Warning] problema ao registrar mesnagem, conteudo vazio")
-            return
-        
-        nova_mensagem = Mensagem(
-            chat_id=id_chat,
-            sessao_id=sessao_id,
-            sender_id=usuario_id,
-            conteudo=conteudo,
-            data_envio=data_de_envio
-        )
-
-        db.session.add(nova_mensagem)
-        db.session.commit()
-
-        return nova_mensagem.id
-
-    except Exception as e:
-        db.session.rollback()
-        print(f"[Error] erro ao persistir mensagem: {e}")
-        return None
-
-
-def registrar_chat(id_usuario, id_materia, primeiro_titulo): 
-    try: 
-        titulo = primeiro_titulo.strip().replace("\n", " ")
-        titulo = " ".join(titulo.split())
-        titulo = titulo[:50]
-
-        if not titulo: 
-            titulo = "Novo chat"
-
-        novo_chat = Chat(
-            aluno_id =id_usuario,
-            materia_id=id_materia,
-            nome=titulo
-        )
-
-        db.session.add(novo_chat)
-        db.session.commit()
-
-        return novo_chat.id
-
-    except Exception as e:
-        db.session.rollback()
-        print(f"[Error] erro ao registrar chat: {e}")
-        return None
